+++ language/org.thingml/src-gen/org/thingml/xthingml/adapters/thingmlmt/thingML/PrimitiveTypeAdapter.java
@@ -28,12 +28,12 @@ public class PrimitiveTypeAdapter extends EObjectAdapter<PrimitiveType> implemen
   }
   
   @Override
-  public int getByteSize() {
+  public long getByteSize() {
     return adaptee.getByteSize();
   }
   
   @Override
-  public void setByteSize(final int o) {
+  public void setByteSize(final long o) {
     adaptee.setByteSize(o);
   }
   
@@ -48,7 +48,7 @@ public class PrimitiveTypeAdapter extends EObjectAdapter<PrimitiveType> implemen
   
   protected final static String NAME_EDEFAULT = null;
   
-  protected final static int BYTE_SIZE_EDEFAULT = 0;
+  protected final static long BYTE_SIZE_EDEFAULT = 0L;
   
   @Override
   public EClass eClass() {
@@ -63,7 +63,7 @@ public class PrimitiveTypeAdapter extends EObjectAdapter<PrimitiveType> implemen
     	case org.thingml.xtext.thingML.ThingMLPackage.PRIMITIVE_TYPE__ANNOTATIONS:
     		return getAnnotations();
     	case org.thingml.xtext.thingML.ThingMLPackage.PRIMITIVE_TYPE__BYTE_SIZE:
-    		return new java.lang.Integer(getByteSize());
+    		return new java.lang.Long(getByteSize());
     }
     
     return super.eGet(featureID, resolve, coreType);
@@ -96,7 +96,7 @@ public class PrimitiveTypeAdapter extends EObjectAdapter<PrimitiveType> implemen
     		getAnnotations().addAll((Collection) newValue);
     		return;
     	case org.thingml.xtext.thingML.ThingMLPackage.PRIMITIVE_TYPE__BYTE_SIZE:
-    		setByteSize(((java.lang.Integer) newValue).intValue());
+    		setByteSize(((java.lang.Long) newValue).longValue());
     		return;
     }
